+++ src/main/java/com/evacipated/cardcrawl/modthespire/lib/SpirePatch.java
@@ -5,12 +5,14 @@ import java.lang.annotation.*;
 @Repeatable(SpirePatches.class)
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface SpirePatch {
+public @interface SpirePatch
+{
     Class<?> clz() default void.class;
     String cls() default "";
     String method();
     Class<?>[] paramtypez() default {void.class};
     String[] paramtypes() default {"DEFAULT"};
+    String requiredModId() default "";
     boolean optional() default false;
 
     String CONSTRUCTOR = "<ctor>";